+++ packages/eslint-plugin-template/src/utils/is-hidden-from-screen-reader.ts
@@ -1,22 +1,23 @@
+import type { TmplAstElement } from '@angular/compiler';
 import { PROPERTY } from './constants';
-import { getLiteralValue } from './get-literal-value';
 import { getAttributeValue } from './get-attribute-value';
+import { getLiteralValue } from './get-literal-value';
 
 /**
- * Returns boolean indicating that the aria-hidden prop
- * is present or the value is true. Will also return true if
- * there is an input with type='hidden'.
- *
- * <div aria-hidden /> is equivalent to the DOM as <div aria-hidden=true />.
+ * Whether an element has the `aria-hidden` property and its value is empty or
+ * `true`. It also returns `true` if the element is an `input` with `type="hidden"`.
  */
-export function isHiddenFromScreenReader(node: any): boolean {
+export function isHiddenFromScreenReader(node: TmplAstElement): boolean {
   if (node.name.toUpperCase() === 'INPUT') {
-    const hidden = getAttributeValue(node, 'type');
-    if (typeof hidden === 'string' && hidden.toUpperCase() === 'HIDDEN') {
+    const typeAttributeValue = getAttributeValue(node, 'type');
+    if (
+      typeof typeAttributeValue === 'string' &&
+      typeAttributeValue.toUpperCase() === 'HIDDEN'
+    ) {
       return true;
     }
   }
 
   const ariaHidden = getLiteralValue(getAttributeValue(node, 'aria-hidden'));
-  return ariaHidden === PROPERTY || ariaHidden === true;
+  return ariaHidden === '' || ariaHidden === PROPERTY || ariaHidden === true;
 }