+++ packages/eslint-plugin-template/tests/rules/accessibility-elements-content.test.ts
@@ -27,31 +27,39 @@ ruleTester.run(RULE_NAME, rule, {
     '<a><app-content></app-content></a>',
     '<a [innerHTML]="dangerouslySetHTML"></a>',
     '<a [innerText]="text"></a>',
+    '<a [outerHTML]="text"></a>',
+    '<a aria-hidden></a>',
+    '<button [attr.aria-hidden]="true"></button>',
+    '<h5 [attr.aria-label]="text"></h5>',
+    '<h6 title="text"></h6>',
   ],
   invalid: [
     convertAnnotatedSourceToFailureCase({
-      messageId,
       description: 'should fail with no content in heading tag',
       annotatedSource: `
         <h1 class="size-1"></h1>
         ~~~~~~~~~~~~~~~~~~~~~~~~
       `,
+      messageId,
+      data: { element: 'h1' },
     }),
     convertAnnotatedSourceToFailureCase({
-      messageId,
       description: 'should fail with no content in anchor tag',
       annotatedSource: `
         <a href="#" [routerLink]="['route1']"></a>
         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
       `,
+      messageId,
+      data: { element: 'a' },
     }),
     convertAnnotatedSourceToFailureCase({
-      messageId,
       description: 'should fail with no content in button tag',
       annotatedSource: `
         <button></button>
         ~~~~~~~~~~~~~~~~~
       `,
+      messageId,
+      data: { element: 'button' },
     }),
   ],
 });